[Refactor] Extract entity queryset and justifications from entity models_django file.
