+++ cosmae/contribution/column/queue/ingest.py
@@ -12,8 +12,9 @@ from cosmae.column.models_django import Column, ColumnHistory
 from cosmae.contribution.column.models_django import ColumnContribution
 from cosmae.contribution.column.queue.util import read_csv_of_candidate
 from cosmae.contribution.models_django import ContributionCandidate
-from cosmae.entity.models_django import Entity, EntityHistory, EntityJustification
+from cosmae.entity.models_django import Entity, EntityHistory
 from cosmae.exception import ColumnExistsException
+from cosmae.justification.models_django import EntityJustification
 from cosmae.merge_request.models_django import ColumnMergeRequest
 from cosmae.util import CosmaeUser, timestamp
 from cosmae.value.models_django import (
