+++ cosmae/models.py
@@ -10,7 +10,8 @@ from cosmae.contribution.column.models_django import (
 from cosmae.contribution.entity.models_django import EntityDuplicate
 from cosmae.contribution.models_django import ContributionCandidate
 from cosmae.edit_session.models_django import EditSession, EditSessionParticipant
-from cosmae.entity.models_django import Entity, EntityJustification
+from cosmae.entity.models_django import Entity
+from cosmae.justification.models_django import EntityJustification
 from cosmae.management.models_django import ConfigValue
 from cosmae.merge_request.entity.models_django import (
     EntityConflictResolution,
