+++ tests/conftest.py
@@ -14,7 +14,8 @@ import tests.column.common as c
 import tests.value.common as cv
 from cosmae.column.models_django import Column, ColumnHistory, column_objects
 from cosmae.edit_session.models_django import EditSession, EditSessionParticipant
-from cosmae.entity.models_django import EntityHistory, EntityJustification
+from cosmae.entity.models_django import EntityHistory
+from cosmae.justification.models_django import EntityJustification
 from cosmae.management.display_txt.util import DISPLAY_TXT_ORDER_CONFIG_KEY
 from cosmae.management.models_django import ConfigValue
 from cosmae.util import CosmaeUser
