+++ tests/contribution/column/queue/ingest_test.py
@@ -10,7 +10,8 @@ from cosmae.column.models_django import Column, ColumnHistory, column_objects
 from cosmae.contribution.column.models_django import ColumnContribution
 from cosmae.contribution.column.queue.ingest import ingest_values_from_csv
 from cosmae.contribution.models_django import ContributionCandidate
-from cosmae.entity.models_django import Entity, EntityJustification
+from cosmae.entity.models_django import Entity
+from cosmae.justification.models_django import EntityJustification
 from cosmae.merge_request.models_django import ColumnMergeRequest
 from cosmae.util import timestamp
 from cosmae.value.models_django import Value
